[MODINVOICE-585] Update order paymentStatus with parameter when paying instead of approving
