+++ src/main/java/org/folio/services/invoice/InvoiceApprovalService.java
@@ -46,7 +46,6 @@ public class InvoiceApprovalService {
   private final InvoiceValidator validator;
   private final InvoiceWorkflowDataHolderBuilder holderBuilder;
   private final PendingPaymentWorkflowService pendingPaymentWorkflowService;
-  private final PoLinePaymentStatusUpdateService poLinePaymentStatusUpdateService;
   private final VendorRetrieveService vendorService;
   private final VoucherCommandService voucherCommandService;
   private final VoucherCreationService voucherCreationService;
@@ -58,7 +57,7 @@ public class InvoiceApprovalService {
       InvoiceFundDistributionService invoiceFundDistributionService, InvoiceLineService invoiceLineService,
       InvoiceValidator validator, InvoiceWorkflowDataHolderBuilder holderBuilder,
       PendingPaymentWorkflowService pendingPaymentWorkflowService,
-      PoLinePaymentStatusUpdateService poLinePaymentStatusUpdateService, VendorRetrieveService vendorService,
+      VendorRetrieveService vendorService,
       VoucherCommandService voucherCommandService, VoucherCreationService voucherCreationService,
       VoucherService voucherService) {
     this.budgetExpenseClassService = budgetExpenseClassService;
@@ -70,7 +69,6 @@ public class InvoiceApprovalService {
     this.validator = validator;
     this.holderBuilder = holderBuilder;
     this.pendingPaymentWorkflowService = pendingPaymentWorkflowService;
-    this.poLinePaymentStatusUpdateService = poLinePaymentStatusUpdateService;
     this.vendorService = vendorService;
     this.voucherCommandService = voucherCommandService;
     this.voucherCreationService = voucherCreationService;
@@ -100,9 +98,7 @@ public class InvoiceApprovalService {
         .map(v -> holders))
       .compose(holders -> budgetExpenseClassService.checkExpenseClasses(holders, requestContext))
       .compose(holders -> pendingPaymentWorkflowService.handlePendingPaymentsCreation(holders, invoice, requestContext))
-      .compose(holders -> poLinePaymentStatusUpdateService.updatePoLinePaymentStatusToApproveInvoice(lines,
-        poLinePaymentStatus, requestContext)
-        .compose(v -> prepareVoucher(invoice, requestContext))
+      .compose(holders -> prepareVoucher(invoice, requestContext)
         .compose(voucher -> updateVoucherWithSystemCurrency(voucher, lines, requestContext))
         .compose(voucher -> voucherCommandService.updateVoucherWithExchangeRate(voucher, invoice, requestContext))
         .compose(voucher -> invoiceFundDistributionService.getAllFundDistributions(lines, invoice, requestContext)
