+++ src/main/java/org/folio/services/invoice/PoLinePaymentStatusUpdateService.java
@@ -1,8 +1,11 @@
 package org.folio.services.invoice;
 
+import com.google.common.annotations.VisibleForTesting;
 import io.vertx.core.Future;
 import io.vertx.core.json.JsonObject;
 import one.util.streamex.StreamEx;
+import org.apache.commons.collections4.CollectionUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.folio.rest.acq.model.orders.CompositePoLine;
@@ -14,17 +17,24 @@ import org.folio.rest.jaxrs.model.InvoiceLine;
 import org.folio.services.order.OrderLineService;
 
 import java.util.ArrayList;
+import java.util.Collections;
+import java.util.EnumSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
+import java.util.Set;
 import java.util.function.Function;
+import java.util.stream.Collectors;
 
 import static io.vertx.core.Future.succeededFuture;
 import static java.util.Collections.emptyList;
 import static java.util.stream.Collectors.groupingBy;
+import static java.util.stream.Collectors.toList;
 import static org.folio.invoices.utils.HelperUtils.collectResultsOnSuccess;
 import static org.folio.invoices.utils.HelperUtils.convertIdsToCqlQuery;
 import static org.folio.rest.RestConstants.MAX_IDS_FOR_GET_RQ;
+import static org.folio.rest.acq.model.orders.CompositePoLine.PaymentStatus.ONGOING;
+import static org.folio.rest.acq.model.orders.CompositePoLine.PaymentStatus.PAYMENT_NOT_REQUIRED;
 import static org.folio.rest.acq.model.orders.PoLine.PaymentStatus.AWAITING_PAYMENT;
 import static org.folio.rest.acq.model.orders.PoLine.PaymentStatus.PARTIALLY_PAID;
 
@@ -35,6 +45,8 @@ public class PoLinePaymentStatusUpdateService {
   private static final String PO_LINE_WITH_ONE_TIME_OPEN_ORDER_QUERY =
     "purchaseOrder.orderType==\"One-Time\" AND purchaseOrder.workflowStatus==\"Open\"";
   private static final String AND = " AND ";
+  private static final Set<CompositePoLine.PaymentStatus> PO_LINE_PAYMENT_IGNORED_STATUSES =
+    EnumSet.of(ONGOING, PAYMENT_NOT_REQUIRED);
 
   private static final Logger logger = LogManager.getLogger();
 
@@ -50,10 +62,10 @@ public class PoLinePaymentStatusUpdateService {
     this.orderLineService = orderLineService;
   }
 
-  public Future<Void> updatePoLinePaymentStatusToApproveInvoice(List<InvoiceLine> invoiceLines, String poLinePaymentStatus,
+  public Future<Void> updatePoLinePaymentStatusToPayInvoice(List<InvoiceLine> invoiceLines, String poLinePaymentStatus,
       RequestContext requestContext) {
     if (poLinePaymentStatus == null) {
-      return succeededFuture();
+      return payPoLines(invoiceLines, requestContext);
     }
     return updatePoLinePaymentStatusUsingParameter(invoiceLines, poLinePaymentStatus, requestContext);
   }
@@ -107,6 +119,81 @@ public class PoLinePaymentStatusUpdateService {
                 invoiceId, requestContext));
           });
       });
+  }
+
+  /**
+   * Updates payment status of the associated PO Lines (when poLinePaymentStatus is not used)
+   *
+   * @param invoiceLines the invoice lines to be paid
+   * @return CompletableFuture that indicates when transition is completed
+   */
+  private Future<Void> payPoLines(List<InvoiceLine> invoiceLines, RequestContext requestContext) {
+    Map<String, List<InvoiceLine>> poLineIdInvoiceLinesMap = groupInvoiceLinesByPoLineId(invoiceLines);
+    return fetchPoLines(poLineIdInvoiceLinesMap, requestContext)
+      .map(this::updatePoLinesPaymentStatus)
+      .compose(poLines -> orderLineService.updateCompositePoLines(poLines, requestContext));
+  }
+
+  private Map<String, List<InvoiceLine>>  groupInvoiceLinesByPoLineId(List<InvoiceLine> invoiceLines) {
+    if (CollectionUtils.isEmpty(invoiceLines)) {
+      return Collections.emptyMap();
+    }
+    return invoiceLines
+      .stream()
+      .filter(invoiceLine -> StringUtils.isNotEmpty(invoiceLine.getPoLineId()))
+      .collect(Collectors.groupingBy(InvoiceLine::getPoLineId));
+  }
+
+  /**
+   * Retrieves PO Lines associated with invoice lines and calculates expected PO Line's payment status
+   * @param poLineIdsWithInvoiceLines map where key is PO Line id and value is list of associated invoice lines
+   * @return map where key is {@link CompositePoLine} and value is expected PO Line's payment status
+   */
+  private Future<Map<CompositePoLine, CompositePoLine.PaymentStatus>> fetchPoLines(Map<String, List<InvoiceLine>> poLineIdsWithInvoiceLines,
+    RequestContext requestContext) {
+    List<Future<CompositePoLine>> futures = poLineIdsWithInvoiceLines.keySet()
+      .stream()
+      .map(poLine -> orderLineService.getPoLine(poLine, requestContext))
+      .collect(toList());
+
+    return collectResultsOnSuccess(futures)
+      .map(compositePoLines ->  compositePoLines.stream()
+        .collect(Collectors.toMap(poLine -> poLine, poLine -> getPoLinePaymentStatus(poLineIdsWithInvoiceLines.get(poLine.getId())))));
+  }
+
+  private List<CompositePoLine> updatePoLinesPaymentStatus(Map<CompositePoLine, CompositePoLine.PaymentStatus> compositePoLinesWithNewStatuses) {
+    return compositePoLinesWithNewStatuses
+      .keySet().stream()
+      .filter(compositePoLine -> isPaymentStatusUpdateRequired(compositePoLinesWithNewStatuses, compositePoLine))
+      .map(compositePoLine -> updatePaymentStatus(compositePoLinesWithNewStatuses, compositePoLine))
+      .collect(toList());
+  }
+
+  @VisibleForTesting
+  boolean isPaymentStatusUpdateRequired(Map<CompositePoLine, CompositePoLine.PaymentStatus> compositePoLinesWithStatus,
+      CompositePoLine compositePoLine) {
+    CompositePoLine.PaymentStatus newPaymentStatus = compositePoLinesWithStatus.get(compositePoLine);
+    return (!newPaymentStatus.equals(compositePoLine.getPaymentStatus()) &&
+      !PO_LINE_PAYMENT_IGNORED_STATUSES.contains(compositePoLine.getPaymentStatus()));
+  }
+
+  private CompositePoLine updatePaymentStatus(Map<CompositePoLine, CompositePoLine.PaymentStatus> compositePoLinesWithStatus,
+      CompositePoLine compositePoLine) {
+    CompositePoLine.PaymentStatus newPaymentStatus = compositePoLinesWithStatus.get(compositePoLine);
+    compositePoLine.setPaymentStatus(newPaymentStatus);
+    return compositePoLine;
+  }
+
+  private CompositePoLine.PaymentStatus getPoLinePaymentStatus(List<InvoiceLine> invoiceLines) {
+    if (isAnyInvoiceLineReleaseEncumbrance(invoiceLines)) {
+      return CompositePoLine.PaymentStatus.FULLY_PAID;
+    } else {
+      return CompositePoLine.PaymentStatus.PARTIALLY_PAID;
+    }
+  }
+
+  private boolean isAnyInvoiceLineReleaseEncumbrance(List<InvoiceLine> invoiceLines) {
+    return invoiceLines.stream().anyMatch(InvoiceLine::getReleaseEncumbrance);
   }
 
   private Future<Void> updatePoLinePaymentStatusUsingParameter(List<InvoiceLine> invoiceLines,
